Restructure domains to onion architecture
