+++ config/services.yaml
@@ -9,8 +9,8 @@ services:
         exclude:
             - '../src/Kernel.php'
 
-    App\ArticleDomain\Repository\ArticleRepositoryInterface: '@App\\Infrastructure\\Repository\\DoctrineArticleRepository'
-    App\UserDomain\Repository\UserRepositoryInterface: '@App\\Infrastructure\\Repository\\DoctrineUserRepository'
+    App\ArticleDomain\Domain\Repository\ArticleRepositoryInterface: '@App\\ArticleDomain\\Infrastructure\\Persistence\\DoctrineArticleRepository'
+    App\UserDomain\Domain\Repository\UserRepositoryInterface: '@App\\UserDomain\\Infrastructure\\Persistence\\DoctrineUserRepository'
     App\SharedKernel\Service\SlugGeneratorInterface: '@App\SharedKernel\Service\SlugGenerator'
     App\SharedKernel\Service\UuidGeneratorInterface: '@App\SharedKernel\Service\UuidGenerator'
     App\SharedKernel\Service\DateTimeProviderInterface: '@App\SharedKernel\Service\DateTimeProvider'
